requirements.txt
cairosvg
defusedxml
django
django-admin-ordering
django-tastypie
emoji
geojson
lxml
pillow
psycopg2
pyyaml
requests
social-auth-app-django
social-auth-core
svgwrite

# tests
behave
behave-django

# misc
django-extensions

# deployment
gunicorn
gevent
